refactor: make room for frameworks in example project
