+++ packages/_example/.env.example
@@ -1,8 +1,14 @@
-FOREST_PREFIX=
-FOREST_AUTH_SECRET=
+FOREST_AGENT_URL=http://localhost:3351
+FOREST_SERVER_URL=https://api.development.forestadmin.com
+
+HTTP_PORT_LOCAL=3351
+
+# Development
 FOREST_AGENT_URL=http://localhost:3351
 FOREST_ENV_SECRET=
-FOREST_SERVER_URL=https://api.forestadmin.com
+FOREST_AUTH_SECRET=
 
-SERVER_PORT=3351
-SERVER_HOST=localhost
+# Production
+# FOREST_AGENT_URL=
+# FOREST_ENV_SECRET=
+# FOREST_AUTH_SECRET=
