+++ packages/_example/src/forest/customizations/rental.ts
