+++ packages/_example/src/forest/datasources/live.ts
@@ -1,10 +1,10 @@
 import LiveDataSource from '@forestadmin/datasource-live/dist/datasource';
 import faker from '@faker-js/faker';
 
-import sequelizeMySql from '../sequelize/mysql';
+import sequelizeMySql from '../../orms/mysql';
 
 // This seed is in this file because the records are loaded in the memory.
-export default async function seed(dataSource: LiveDataSource) {
+export async function seedLiveDatasource(dataSource: LiveDataSource) {
   try {
     const storeRecords = await sequelizeMySql.model('store').findAll({ attributes: ['id'] });
     const addressRecords = storeRecords.map((storeRecord, index) => ({
@@ -19,3 +19,12 @@ export default async function seed(dataSource: LiveDataSource) {
     console.log('Please run `yarn db:up & yarn db:seed` to run the db with data');
   }
 }
+
+export const liveDatasourceSchema = {
+  address: {
+    id: { columnType: 'Number', isPrimaryKey: true, type: 'Column' },
+    zipCode: { columnType: 'String', type: 'Column' },
+    address: { columnType: 'String', type: 'Column' },
+    storeId: { columnType: 'Number', type: 'Column' },
+  },
+} as const;
